backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.2
python-multipart==0.0.6
kafka-python==2.0.2
